+++ src/app/shared/module/poe/service/item/item-exchange-rate.service.spec.ts
@@ -36,7 +36,7 @@ describe('ItemExchangeRateService', () => {
 
     currencyService
       .searchById('chaos')
-      .pipe(flatMap((chaos) => sut.get(item, [chaos])))
+      .pipe(flatMap((chaos) => sut.get(item, [chaos], false)))
       .subscribe(
         (result) => {
           expect(result).toBeTruthy()