+++ src/app/shared/module/poe/type/item.type.ts
@@ -120,6 +120,7 @@ export enum ItemCategory {
   SanctumRelic = 'sanctum.relic',
   Tincture = "tincture",
   Corpse = "corpse",
+  NecropolisPack = "necropolis.pack",
   Idol = "idol",
   Graft = "graft",
   Wombgift = "wombgift",
@@ -133,7 +134,6 @@ export enum ItemCategory {
   CurrencyTattoo = 'currency.tattoo',
 
   // Deprecated (due to previous league mechanics not going core or being removed)
-  NecropolisPack = "necropolis.pack",
   AzmeriTincture = "azmeri.tincture",
   AzmeriCharm = "azmeri.charm",
   AzmeriCorpse = "azmeri.corpse",